+++ api_finalproject/Models/Pedido.cs
@@ -8,8 +8,8 @@ namespace api_finalproject.Models
     public partial class Pedido
     {
         public int Id { get; set; }
-        public int Latitud { get; set; }
-        public int Longitud { get; set; }
+        public float Latitud { get; set; }
+        public float Longitud { get; set; }
         public int OrdenId { get; set; }
         public int EstadoPedidoId { get; set; }
